# backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users/me": {
      "patch": {
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update current user profile",
        "tags": [
          "Users"
        ]
      },
      "get": {
        "operationId": "UsersController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get current user profile",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/properties/search": {
      "get": {
        "operationId": "PropertiesController_search",
        "parameters": [
          {
            "name": "city",
            "required": false,
            "in": "query",
            "schema": {
              "example": "Ho Chi Minh City",
              "type": "string"
            }
          },
          {
            "name": "star_rating",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "example": 4,
              "type": "number"
            }
          },
          {
            "name": "min_price",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "example": 500000,
              "type": "number"
            }
          },
          {
            "name": "max_price",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "example": 2500000,
              "type": "number"
            }
          },
          {
            "name": "check_in",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-06-01",
              "type": "string"
            }
          },
          {
            "name": "check_out",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-06-03",
              "type": "string"
            }
          },
          {
            "name": "rooms_needed",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns active properties matching search and availability."
          }
        },
        "summary": "Search public properties",
        "tags": [
          "Properties"
        ]
      }
    },
    "/api/properties/{slug}": {
      "get": {
        "operationId": "PropertiesController_findBySlug",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-06-01",
              "type": "string"
            }
          },
          {
            "name": "check_out",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-06-03",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns property details and room types."
          }
        },
        "summary": "Get public property detail by slug",
        "tags": [
          "Properties"
        ]
      }
    },
    "/api/properties/id/{id}": {
      "get": {
        "operationId": "PropertiesController_findById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "check_in",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-06-01",
              "type": "string"
            }
          },
          {
            "name": "check_out",
            "required": false,
            "in": "query",
            "schema": {
              "example": "2026-06-03",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns property details and room types."
          }
        },
        "summary": "Get public property detail by ID",
        "tags": [
          "Properties"
        ]
      }
    },
...
    "/api/reviews": {
      "post": {
        "operationId": "ReviewsController_createReview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReviewBodyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted successfully."
          },
          "400": {
            "description": "Booking is not eligible for a review."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit a post-stay review for a completed booking",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/api/reviews/property/{id}": {
      "get": {
        "operationId": "ReviewsController_findByProperty",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns property reviews."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all reviews for a property",
        "tags": [
          "Reviews"
        ]
      }
    },
...
    "/api/payments/checkout/{bookingId}": {
      "post": {
        "operationId": "PaymentsController_checkout",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Checkout session created for the authenticated customer."
          },
          "401": {
            "description": "Missing or invalid access token."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a payment checkout session for a booking",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/process": {
      "post": {
        "operationId": "PaymentsController_processPayment",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Payment processed successfully."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Simulate successful payment processing",
        "tags": [
          "Payments"
        ]
      }
    },
    "/api/payments/webhook": {
      "post": {
        "operationId": "PaymentsController_handleWebhook",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully."
          }
        },
        "summary": "Receive payment provider webhook notification",
        "tags": [
          "Payments"
        ]
      }
    }
  },
  "info": {
    "title": "NoWayHome API",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Nguyễn Văn B",
            "description": "Full name of the user"
          },
          "phone": {
            "type": "string",
            "example": "0123456789",
            "description": "Phone number"
          },
          "dateOfBirth": {
            "type": "string",
            "example": "01-01-1990",
            "description": "Date of birth (DD-MM-YYYY format)"
          },
          "gender": {
            "type": "string",
            "enum": [
              "male",
              "female",
              "other"
            ],
            "example": "male",
            "description": "Gender"
          }
        }
      },
...
      "CreateReviewBodyDto": {
        "type": "object",
        "properties": {
          "bookingId": {
            "type": "string",
            "example": "1",
            "description": "ID of the booking to review"
          },
          "rating": {
            "type": "number",
            "example": 4.5,
            "description": "Overall rating from 1 to 5"
          },
          "comment": {
            "type": "string",
            "example": "Great stay!",
            "description": "Optional review comment"
          }
        },
        "required": [
          "bookingId",
          "rating"
        ]
      },
      "WebhookDto": {
        "type": "object",
        "properties": {
          "bookingCode": {
            "type": "string",
            "example": "NWH-20260518-0001",
            "description": "Booking code returned by the payment provider."
          },
          "transactionStatus": {
            "type": "string",
            "enum": [
              "SUCCESS",
              "FAILED"
            ],
            "example": "SUCCESS",
            "description": "Final transaction status reported by the payment provider."
          }
        },
        "required": [
          "bookingCode",
          "transactionStatus"
        ]
      }
    }
  }
}